refactor: move refs operation from ipfs-http
